#feat #map

+++ include/core/common.h
@@ -10,7 +10,7 @@ typedef void *(*neo_alloc_fn_t)(size_t size);
 
 typedef void (*neo_free_fn_t)(void *ptr);
 
-typedef int8_t (*neo_compare_fn_t)(const void *, const void *, void *);
+typedef int32_t (*neo_compare_fn_t)(const void *, const void *, void *);
 
 typedef uint32_t (*neo_hash_fn_t)(const void *, uint32_t, void *);
